// integrations/netsuite-tba/schema.ts
export const netsuiteInvoiceCreateInputSchema = z.object({
    customerId: z.string().min(1),
    currency: z.string().min(1),
    description: z.string().optional(),
    status: z.string().min(1),
    lines: z.array(
        z.object({
            itemId: z.string().min(1),
            quantity: z.number().min(0),
            locationId: z.string().min(1),
            amount: z.number().min(0),
            vatCode: z.string().optional(),
            description: z.string().optional()
        })
    )
});